Refactor out sendAccountRemovedToSegment
